Backwork Completed along with Spring Security

+++ School-Students/src/main/java/Student_Details/School/Students/Repo/Teacher_Repo.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Map;
 
-public interface Teacher_Repo extends JpaRepository<Teachers, Long> {
+public interface Teacher_Repo extends JpaRepository<Teachers, Integer> {
     boolean existsByUserName(String userName);
 
     Teachers findByUserName(String userName);
